+++ src/clova/clova.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ClovaController } from './clova.controller';
 import { ClovaService } from './clova.service';
 import { FeedbackService } from './feedback.service';
-import { ParsedSentenceService } from './parsedSentence.service';
+import { ParsedSentenceService } from './parsed-sentence.service';
 import { PartialModificationService } from './partial-modification.service';
 
 @Module({
